Rename initJsonStorage to initSettingsStorage
