+++ main-process/fragmemoSettings/initSettingsStorage.ts
